feat(core): move constants
