+++ packages/core/src/constants/container-module.ts
