+++ packages/core/src/index.ts
@@ -3,5 +3,6 @@ export * from './application/app-service/service';
 export * from './navigation/route/service';
 export * from './navigation/matcher/service';
 export * from './navigation/router/service';
-export * from './constants';
+export * from './constants/container-module';
+export * from './constants/status';
 export * from './provider';
