+++ packages/core/src/provider/index.ts
@@ -1,4 +1,4 @@
-import { VERSEA_METADATA_PROVIDE_KEY } from '../constants';
+import { VERSEA_METADATA_PROVIDE_KEY } from '../constants/container-module';
 import { createProvider } from './creator';
 
 const { provide, provideValue, buildProviderModule } = createProvider(VERSEA_METADATA_PROVIDE_KEY);
